pyproject.toml
[tool.poetry]
name = "angrymetalguy-to-spotify"
version = "0.1.0"
description = "Parses \"Angry Metal Guy\" reviews and auto publishes albums with high scores to a spotify playlist"
authors = ["Mauricio R. Cruz <mauriciorcruz@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "angrymetalguy_to_spotify"}]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.2"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
spotipy = "^2.23.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.0"
isort = "^5.12.0"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
float_to_top = "true"
profile = "black"

[tool.black]
line-length = 130

[tool.mypy]
python_version = "3.10"
warn_return_any = true
disallow_untyped_defs = true
